amd/addrlib: fix DEBUG guards
Use the normal `#ifdef` style used by the rest of Mesa, instead of
sometimes locally defining DEBUG on non-debug builds and then relying on
the precompiler converting #undef'ed macros to 0 when checking.

Signed-off-by: Eric Engestrom <eric.engestrom@imgtec.com>
---
I would argue this block should go, as all it does is enable debug
prints on release builds with asserts. It doesn't make sense to me that
the fact asserts are enabled is relevant to this.