u_debug: do not compile asserts when they are disabled
Commit f0ba7d897d1c22202531a added this code to expose asserts to the
compiler in an attempt to hide 'unused variable' warnings, incorrectly
claiming it was a no-op. After this change, the whole content of the
assert must be understandable by the compiler, which isn't true if
variable or members are correctly guarded by NDEBUG.

Fix this by effectively reverting f0ba7d897d1c22202531a.

Unused variables warnings can be addressed by marking the variables as
MAYBE_UNUSED instead, as is done in the rest of Mesa.

Fixes: f0ba7d897d1c22202531a "util: better fix for unused variable
       warnings with asserts"
Cc: Keith Whitwell <keithw@vmware.com>
Reported-by: Gert Wollny <gw.fossdev@gmail.com>
Signed-off-by: Eric Engestrom <eric.engestrom@imgtec.com>