docs: add EGL to features.txt
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=104478
Cc: Dennis Schridde <dennis.schridde@uni-heidelberg.de>
Cc: Romain Failliot <romain.failliot@foolstep.com> # MesaMatrix maintainer
Signed-off-by: Eric Engestrom <eric.engestrom@imgtec.com>
---
The list only contains supported extensions; do we want to add the 100+
unsupported ones as well? Or only add the ones we want to see supported
soon?

Romain, would you be able to parse this format? It's a bit more complex
than GL/VK extensions, as they depend on two factors: the driver, and
the EGL platform. In the case of the X11 platform, it can even depend on
the protocol used by the X server, but I'm happy to remove this
precision.